+++ rollbar-cli/ChangeLog.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.2.0] - 2026-08-18
+
+### Changed
+
+- Version bump only, keeping in step with the other rollbar packages; the CLI
+  does not report exceptions.
+
 ## [1.1.0] - 2024-05-28
 
 ### Changed